pyproject.toml
[tool.poetry]
name = "python-rn"
version = "0.1.0"
description = ""
authors = ["Bentroen <bemcdc@gmail.com>"]
readme = "README.md"
packages = [{include = "python_rn"}]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.5.0"
torch = "^2.0.1"
pandas = "^2.0.1"
ipykernel = "^6.23.1"
matplotlib = "^3.7.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
